requirements.txt
blinker==1.9.0
click==8.1.8
coloredlogs==15.0.1
filelock==3.18.0
Flask==3.1.0
flatbuffers==25.2.10
fsspec==2025.3.2
humanfriendly==10.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.4
onnxruntime==1.21.0
opencv-python-headless==4.11.0.86
packaging==24.2
pillow==11.1.0
progressbar2==4.5.0
protobuf==6.30.2
python-dotenv==1.1.0
python-utils==3.9.1
setuptools==78.1.0
sympy==1.13.1
torch==2.6.0+cpu
torchvision==0.21.0+cpu
triton==3.2.0
typing_extensions==4.13.0
Werkzeug==3.1.3
